算法/LLM_APPLICATION/mcp/requirements.txt
mcp>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langsmith>=0.1.0
python-dotenv>=1.0.0
asyncio
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.8.0 